Move wrapper width calculator to separate file
